+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Updated `make_lossless!` to make the model truly lossless, by including all line parameters
 - Added `adjust_transformer_limits!` to multiple transformer variable bounds
 - Added transformations `adjust_small_line_impedances!`, `adjust_small_line_admittances!`, `adjust_small_line_lengths!`, to help prune bad eng-model data
 - Fixed `set_time_elapsed!` multinetwork function to update `"time"` and `"mn_lookup"`